Get rid of reinterpret cast. Fix clock to terminate.
NOTE: What do we do if clock is stopped? The
channel endpoints to the clock are still held by
fibres. In addition, some may be waiting on a timeout.
So the current solution is temporary and unsafe.
The rule is: don't stop the clock until there will
all timeouts, including any future alarm requests,
have been satisfied. This should be made safer.
The design sucks. Note the mainline CANNOT terminate
until the clock service thread is joined.