Adjust passing of parameters.
Continuations now require the containing fibre in their
constructor. For a spawn, this is not known so a nullptr
must be passed instead, the system sets it after constructing
the fibre. For a call, the fibre is passed by the caller.

NOTE: In the case of a spawn, the user typically creates the
continuation and then calls it and then passes THAT to the
system via an SVC. Only then is the fibre set. This means
the call method MUST NOT access the fibre non-static member
(because it is a nullptr until after the SVC). in particular,
since global is a member of fibre now, and the allocator
is a member of global, the allocator cannot be used explicitly in a call.

Call is ONLY allowed to assign arguments to parameters.
Note, if the assignment invokes an allocator via C++ standard
library that use is OK.

This is all very messy.  In the abstract, fibres now exclusively
own continuations, the latter have a weak pointer to the fibre.
Although continuations now cannot move to another fibre they
might be orhphaned. In this case the fibre could be deleted
whilst the continuation still exists. This must not be allowed
to happen. In particular, continuations are deleted if they live
on a channel which becomes unreachable, and that WILL certainly
invoke the fibre's allocator to perform the deletion. I think the
exiting code is correct (but it is recursive so who knows .. :)