Allocators are now ref counted.
The type alloc_ref_t is a strong (ref counted) smart pointer
to an allocator. It is passed around and ensure the continued
existence of an allocator whilst required by ensuring it exists
when reachable, and deleted when not.

As an exception channel endpoints use weak pointers to the
underlying allocator. The allocator used must be the system
allocator and therefore a weak pointer suffices provided
the endpoint lifetime is included in the system object lifetime.
This should always be.